bug_fix_3858424
Issue

-> The uint8_t bus of getFruInfo() restricted the number of logical I2C
   buses to 255 and also static_cast'ed to uint8_t.

-> Whenever I2C buses crosses 256 bus limit because of type casting to
   uint_8 the bus becomes 0 and which creates coredump while accessing it.

For reference:
https://lore.kernel.org/all/CACWQX82=MuAavxCqOerxi-Sdywh0xatb-f+1YzGyVSg74oNGqA@mail.gmail.com/T/

Fixes:-
changed type casting to uint16_t instead of uint8_t .

Testing:-
1. Fru listing after and before of power off , power on and reboot

Fixes nvbug https://nvbugswb.nvidia.com/NvBugs5/SWBug.aspx?bugid=3858424&cmtNo=

+++ src/FruDevice.cpp
@@ -682,9 +682,9 @@ void addFruObjectToDbus(
             if (std::regex_match(path, std::regex(productName + "(_\\d+|)$")))
             {
                 if (isMuxBus(bus) && address == busIface.first.second &&
-                    (getFRUInfo(static_cast<uint8_t>(busIface.first.first),
+                    (getFRUInfo(static_cast<uint16_t>(busIface.first.first),
                                 static_cast<uint8_t>(busIface.first.second)) ==
-                     getFRUInfo(static_cast<uint8_t>(bus),
+                     getFRUInfo(static_cast<uint16_t>(bus),
                                 static_cast<uint8_t>(address))))
                 {
                     // This device is already added to the lower numbered bus,
@@ -913,7 +913,7 @@ bool writeFRU(uint8_t bus, uint8_t address, const std::vector<uint8_t>& fru)
 }
 
 void rescanOneBus(
-    BusMap& busmap, uint8_t busNum,
+    BusMap& busmap, uint16_t busNum,
     boost::container::flat_map<
         std::pair<size_t, size_t>,
         std::shared_ptr<sdbusplus::asio::dbus_interface>>& dbusInterfaceMap,
@@ -935,7 +935,7 @@ void rescanOneBus(
     {
         if (dbusCall)
         {
-            std::cerr << "Unable to access i2c bus " << static_cast<int>(busNum)
+            std::cerr << "Unable to access i2c bus " << static_cast<uint16_t>(busNum)
                       << "\n";
             throw std::invalid_argument("Invalid Bus.");
         }
@@ -1094,7 +1094,7 @@ bool updateFRUProperty(
     std::vector<uint8_t> fruData;
     try
     {
-        fruData = getFRUInfo(static_cast<uint8_t>(bus),
+        fruData = getFRUInfo(static_cast<uint16_t>(bus),
                              static_cast<uint8_t>(address));
     }
     catch (const std::invalid_argument& e)
@@ -1382,14 +1382,14 @@ int main()
                      objServer, systemBus);
     });
 
-    iface->register_method("ReScanBus", [&](uint8_t bus) {
+    iface->register_method("ReScanBus", [&](uint16_t bus) {
         rescanOneBus(busMap, bus, dbusInterfaceMap, true, unknownBusObjectCount,
                      powerIsOn, objServer, systemBus);
     });
 
     iface->register_method("GetRawFru", getFRUInfo);
 
-    iface->register_method("WriteFru", [&](const uint8_t bus,
+    iface->register_method("WriteFru", [&](const uint16_t bus,
                                            const uint8_t address,
                                            const std::vector<uint8_t>& data) {
         if (!writeFRU(bus, address, data))
@@ -1467,7 +1467,7 @@ int main()
                                           << "\n";
                                 continue;
                             }
-                            rescanOneBus(busMap, static_cast<uint8_t>(bus),
+                            rescanOneBus(busMap, static_cast<uint16_t>(bus),
                                          dbusInterfaceMap, false,
                                          unknownBusObjectCount, powerIsOn,
                                          objServer, systemBus);
